requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.7
click==8.1.7
colorclass==2.2.2
coloredlogs==15.0.1
compressed-rtf==1.0.6
cryptography==41.0.4
dataclasses-json==0.5.14
diskcache==5.6.3
easygui==0.98.3
ebcdic==1.1.1
emoji==2.8.0
extract-msg==0.45.0
fastapi==0.99.1
filelock==3.12.4
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
gpt4all==1.0.8
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
IMAPClient==2.3.1
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.274
langsmith==0.0.40
lark==1.1.5
llama-cpp-python==0.1.81
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
msoffcrypto-tool==5.1.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.26.0
olefile==0.46
oletools==0.60.1
onnxruntime==1.16.0
overrides==7.4.0
packaging==23.1
pandoc==2.3
pcodedmp==1.2.6
pika==1.3.2
Pillow==10.0.1
plumbum==1.8.2
ply==3.11
posthog==3.0.2
protobuf==4.24.3
pulsar-client==3.3.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
PyMuPDF==1.23.1
PyMuPDFb==1.23.0
pypandoc==1.11
pyparsing==2.4.7
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
PyYAML==6.0.1
red-black-tree-mod==1.20
regex==2023.8.8
requests==2.31.0
RTFDE==0.1.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzlocal==5.0.1
unstructured==0.10.8
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
